Revert "TensorFlow: upstream changes from the afternoon."
This reverts commit f2102f4e2c1c87f1d1bf9ab856a2849c54478760.

+++ tensorflow/python/summary/event_accumulator.py
@@ -1,7 +1,4 @@
 """Takes a generator of values, and accumulates them for a frontend."""
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 import collections
 import threading
@@ -396,7 +393,7 @@ class EventAccumulator(object):
     bucket_limit = list(histo.bucket_limit)
 
     bucket_total = sum(bucket)
-    fraction_weights = [10000 * x / bucket_total for x in bucket]
+    fraction_weights = [float(10000*x)/bucket_total for x in bucket]
     cumsum_weights = _CumulativeSum(fraction_weights)
 
     percentiles = [
