+++ tensorflow/python/training/summary_io.py
@@ -1,9 +1,5 @@
 """Reads Summaries from and writes Summaries to event files."""
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import os.path
 import Queue
 import threading
@@ -109,7 +105,7 @@ class SummaryWriter(object):
       summary = summ
     event = event_pb2.Event(wall_time=time.time(), summary=summary)
     if global_step is not None:
-      event.step = int(global_step)
+      event.step = long(global_step)
     self.add_event(event)
 
   def add_event(self, event):
@@ -133,7 +129,7 @@ class SummaryWriter(object):
     """
     event = event_pb2.Event(wall_time=time.time(), graph_def=graph_def)
     if global_step is not None:
-      event.step = int(global_step)
+      event.step = long(global_step)
     self._event_queue.put(event)
 
   def flush(self):
